IMPALA-14066 (Part 2): Fixed rebased Kudu source code to compile
Based on Wehzhe Zhou's commit 92ce6fe, which fixed the code after the
previous rebase. Also accounted for new changes.

Reviewed-on: http://gerrit.cloudera.org:8080/22969
Tested-by: Daniel Becker <daniel.becker@cloudera.com>

+++ be/src/kudu/util/env_posix.cc
@@ -461,7 +461,7 @@ Status DoEncryptV(const EncryptionHeader* eh,
   const auto* cipher = GetEVPCipher(eh->algorithm);
   if (!cipher) {
     return Status::RuntimeError(
-        StringPrintf("no cipher for algorithm 0x%02x", eh->algorithm));
+        StringPrintf("no cipher for algorithm 0x%02x", (unsigned int) eh->algorithm));
   }
   auto ctx = ssl_make_unique(EVP_CIPHER_CTX_new());
   OPENSSL_RET_IF_NULL(ctx, "failed to create cipher context");
@@ -510,7 +510,7 @@ Status DoDecryptV(const EncryptionHeader* eh, uint64_t offset, ArrayView<Slice> 
   const auto* cipher = GetEVPCipher(eh->algorithm);
   if (!cipher) {
     return Status::RuntimeError(
-        StringPrintf("no cipher for algorithm 0x%02x", eh->algorithm));
+        StringPrintf("no cipher for algorithm 0x%02x", (unsigned int) eh->algorithm));
   }
   auto ctx = ssl_make_unique(EVP_CIPHER_CTX_new());
   OPENSSL_RET_IF_NULL(ctx, "failed to create cipher context");
